correct auth method and remove list_tool file.
